Major rework of wiring
This was motivated by the need to be able to wire a part's output
to multiple outputs of a chip (which was impossible so far)

- Chip() now builds a wiring tree.
- Added "Pinout" field in *PartSpec. In simple cases, it just maps input
and output names to themselves. However, if a PartSpec wires multiple
outputs together, it must create a single internal pin name and map
those outputs to that internal pin name in the Pinout field. The
Component function must then use that internal pin name.